Fixed bug with find draft returning multiple elements
System now allows multiple drafts to be returned when listing drafts by user and draft type, in that case the latest one is returned. Before creating new draft, existing ones with the same userId and type are deleted. For future considerations there should be unique index added to draft with usedId + type fields. This should be done after a while so system has repaired itself.